google-sheets-sync/requirements.txt
python-jobspy
google-auth
google-api-python-client
python-dotenv
pandas
pydantic
